Auto-remove coverage files in TestTools module.
When jltest is used with coverage enabled, coverage data is also
generated within the TestTools module. This commit adds support for
automatic cleanup of coverage files left over from previous uses
of jltest. The coverage data from the last use of jltest (with
coverage enabled) is not removed, but this is desireable in order
to support assessment of coverage of the TestTools codebase during
development and testing of the TestTools package itself.